+++ ApiGateway/src/main/java/com/ApiGateway/LoggingFilter.java
@@ -16,7 +16,7 @@ public class LoggingFilter implements GlobalFilter{
 	
 	@Override
 	public Mono<Void> filter(ServerWebExchange exchange, GatewayFilterChain chain) {
-		logger.info("Path ,{} "+exchange.getRequest().getPath());
+		logger.info("Path ,{} ",exchange.getRequest().getPath());
 		return chain.filter(exchange);
 	}
