+++ ApiGateway/src/main/resources/application.properties
@@ -5,4 +5,5 @@ eureka.client.serviceUrl.defaultZone=http://localhost:8761/eureka
 spring.cloud.gateway.discovery.locator.enabled=true
 spring.cloud.gateway.discovery.locator.lower-case-service-id=true
 spring.sleuth.sampler.probability=1.0
-spring.config.import=optional:configserver:http://localhost:8888
+eureka.instance.hostname=localhost
+spring.cloud.discovery.enabled=true
