+++ SpringApp1/src/main/java/com/SpringApp1/Controller/SpringAppOneEndPoints.java
@@ -1,15 +1,15 @@
-package com.SpringApp2.controller;
+package com.SpringApp1.Controller;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/springapp2")
-public class Endpoint2 {
+@RequestMapping("/helloworld")
+public class SpringAppOneEndPoints {
 
-	@GetMapping("/hello2")
-	public String hello2() {
-		return "Spring App 2 Controller";
+	@GetMapping("/hello1")
+	public String hello1() {
+		return "Spring App 1 Controller";
 	}
 }
