+++ SpringApp1/src/main/resources/application.properties
@@ -2,3 +2,4 @@ spring.config.import=optional:configserver:http://localhost:8888
 server.port=8081
 spring.application.name=SpringApp1
 eureka.client.service-url.defaultZone=http://localhost:8761/eureka
+eureka.instance.hostname=localhost
