+++ SpringApp2/src/main/java/com/SpringApp2/controller/SpringAppTwoEndPoints.java
@@ -1,15 +1,15 @@
-package com.SpringApp1.Controller;
+package com.SpringApp2.controller;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/springapp1")
-public class Endpoint1 {
+@RequestMapping("/helloworld")
+public class SpringAppTwoEndPoints {
 
-	@GetMapping("/hello1")
-	public String hello1() {
-		return "Spring App 1 Controller";
+	@GetMapping("/hello2")
+	public String hello2() {
+		return "Spring App 2 Controller";
 	}
 }
